test: split provider endpoint tests
